updated docs of avatar size: https://discord.com/developers/docs/reference#image-formatting-image-base-url

+++ include/dpp/guild.h
@@ -214,10 +214,10 @@ public:
 	/**
 	 * @brief Returns the members's per guild avatar if they have one, otherwise returns an empty string
 	 * 
-	 * @param size The size of the avatar in pixels. If not specified, the default sized avatar is returned.
+	 * @param size The size of the avatar in pixels. It can be any power of two between 16 and 4096. If not specified, the default sized avatar is returned.
 	 * @return std::string avatar or empty string
 	 */
-	std::string get_avatar_url(uint32_t size = 0) const;
+	std::string get_avatar_url(uint16_t size = 0) const;
 
 	/**
 	 * @brief Set the nickname
